MVVM refactoring
