+++ pISSStream/AppDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  AppDelegate.swift
+//  pISSStream
+//
+//  Created by durul dalkanat on 12/25/24.
+//
+
+import SwiftUI
+
+class AppDelegate: NSObject, NSApplicationDelegate {
+    func applicationDidFinishLaunching(_ aNotification: Notification) {
+        NSApp.setActivationPolicy(.accessory)
+    }
+}
