+++ pISSStream/View/PissLabel.swift
@@ -1,0 +1,17 @@
+//
+//  PissLabel.swift
+//  pISSStream
+//
+//  Created by durul dalkanat on 12/25/24.
+//
+
+import SwiftUI
+
+struct PissLabel: View {
+    let amount: String
+
+    var body: some View {
+        Text("🧑🏽‍🚀🚽\(amount)")
+            .font(.system(size: 12, weight: .bold, design: .default))
+    }
+}
